Make Stockfish::PSQT::psq constexpr.
Currently the anonymous arrays PBonus and Bonus are stored in .rodata
(nm stockfish | grep Bonus) and the array PSQT::psq is stored in .bss and
initialized with PSQT::init().
This patch makes Stockfish::PSQT::psq constexpr, its values are computed
at compile time and stored directly in .rodata, and the anonymous arrays
are not stored because they are used only when compiling.

Bench: 4321677

No functional change